refactor: SpaceListItem -> MapListItem
