+++ frontend/src/components/MapListItem/MapListItem.styles.ts
